Adaption to new skm webquery.
Token can be collected from the website in webquery.
Few edits still missing wrt. date and number format in case we wish to be able to change output. Not critical so I will leave it out for now.

At the current moment there is an issue with output only showing two decimals (it was 4 in the old one). SYSPower will return with an answer.